Fix a shift warning.
We only need 32-bits for this field, the shift is correct.

+++ Core/MIPS/IR.h
@@ -73,7 +73,7 @@ struct IREntry {
 
 	// Clobbered state. Can discard registers marked as clobbered later.
 	u64 clobberedGPR;
-	u64 clobberedFPR;
+	u32 clobberedFPR;
 
 	// u32 liveVPR[4];  // TODO: For now we assume all VPRs are live at all times.
